sdm845-common: prop: Disable backpressure propagation for GL client
Previously we relied on debug.sf.disable_backpressure=1 to disable
backpressure propagation. Now that the prop is gone in S, drop
the backpressure prop for GL client to maintain the same behavior
as before. From past experience, backpressure propagation often
increases jank rate instead of reducing it for some reason.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>